+++ src/main/java/db/DataBase.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 import com.google.common.collect.Maps;
 
-import model.User;
+import mvc.model.User;
 
 public class DataBase {
     private static Map<String, User> users = Maps.newHashMap();
