--- src/main/java/http/request/ProtocolType.java
@@ -1,5 +1,0 @@
-package http.request;
-
-public enum ProtocolType {
-    HTTP
-}
