+++ src/main/java/http/request/protocol/ProtocolType.java
@@ -1,0 +1,5 @@
+package http.request.protocol;
+
+public enum ProtocolType {
+    HTTP, HTTPS
+}
