+++ src/main/java/mvc/model/User.java
@@ -1,4 +1,4 @@
-package model;
+package mvc.model;
 
 public class User {
     private String userId;
