+++ src/test/java/webserver/request/RequestBodyTest.java
@@ -1,5 +1,6 @@
 package webserver.request;
 
+import http.request.RequestBody;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
